test: fix jest tests after ad63b5eb34c
